Finished message protocol format specification
Added all the message protocal format specification to Message.
Decoding message as specified in the protocol to internal data
representation is done up to actions, converting state message
still needs to finish a method in game builder and add a few
setter functions to the state interface, still need to implement
all the message to convert from internal representation to messages.